tranlocation funcs in module readers
