+++ tests/test_utils.py
@@ -1,10 +1,10 @@
 import json
 from unittest.mock import patch
 
-from src.utils import get_transaction_info
+from src.utils import get_transaction_info_from_json
 
 
-def test_get_transaction_info():
+def test_get_transaction_info_from_json():
     """Проверка работы функции с имитацией открытия JSON-файла"""
     mock_data = [
         {
@@ -20,7 +20,7 @@ def test_get_transaction_info():
     with patch("builtins.open") as mock_open:
         mock_file = mock_open.return_value.__enter__.return_value
         mock_file.read.return_value = json.dumps(mock_data)
-        assert get_transaction_info("../data/operations.json") == [
+        assert get_transaction_info_from_json("../data/operations.json") == [
             {
                 "id": 123,
                 "state": "EXECUTED",
@@ -33,15 +33,15 @@ def test_get_transaction_info():
         ]
 
 
-def test_get_transaction_info_invalid_file(capsys):
+def test_get_transaction_info_json_invalid_file(capsys):
     """Проверка работы функции, когда указан несуществующий файл"""
-    get_transaction_info(filename="../data/kakayato_hren.json")
+    get_transaction_info_from_json(filename="../data/kakayato_hren.json")
     captured = capsys.readouterr()
     assert captured.out == ""
 
 
-def test_get_transaction_info_no_file(capsys):
+def test_get_transaction_info_json_no_file(capsys):
     """Проверка работы функции, когда файла нет"""
-    get_transaction_info(filename="")
+    get_transaction_info_from_json(filename="")
     captured = capsys.readouterr()
     assert captured.out == ""
